requirements.txt
ollama>=0.1.6
pydantic>=2.0.0
rich>=13.0.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
requests>=2.31.0
chromadb>=0.4.20
faster-whisper>=1.0.0
youtube-transcript-api>=0.6.0
psutil>=5.9.0
beautifulsoup4>=4.12.0
python-whois>=0.8.0
PyPDF2>=3.0.0
feedparser>=6.0.0
Faker>=20.0.0
gTTS>=2.4.0
dnspython>=2.4.0
apscheduler>=3.10.0
markdown>=3.5.0
pandas>=2.0.0
ta>=0.11.0
yfinance>=0.2.30
playwright>=1.40.0
